Rename "RenderUtils" to "WaypointRenderer"
